Implement custom connection status hook
